refactor: clean up WIR backend post-cutover (4 tasks)
- Remove debug file writes (/tmp/wir_debug_core.wat) from tir_to_wir
- Merge wasm_plan pipeline step into tir_to_wir::compile_with_wir;
  type utilities move to tir_to_wir/types.rs, planning logic to
  tir_to_wir::plan_project(). Pipeline: optimize -> tir_to_wir -> wir_emit
- Delete copy_context.rs; move expand_copy_types into optimize_rewrite.rs
- Convert tir_to_wir to 2018 module style (tir_to_wir.rs + tir_to_wir/)

All 3380 E2E tests pass.
