+++ wado-compiler/src/tir_to_wir.rs
@@ -2,9 +2,9 @@
 //! then emits it as a Wasm component binary.
 //!
 //! This module is the entry point for the WIR pipeline:
-//!   `compile_with_wir(&Project) -> Vec<u8>`
+//!   `compile_with_wir(Project) -> Vec<u8>`
 //!
-//! The pipeline: Project → `build_wir_module` (`WirModule`) → `emit` (Wasm bytes)
+//! The pipeline: Project → planning → `build_wir_module` (`WirModule`) → `emit` (Wasm bytes)
 
 use crate::project::Project;
 use crate::wir::WirModule;
@@ -16,13 +16,28 @@ mod functions;
 mod translate;
 mod types;
 
+/// Run the planning phase — previously the standalone `wasm_plan` pipeline step.
+///
+/// Sets `project.has_http_handler_export` from world analysis and
+/// populates `project.component_plan` for use by `build_wir_module`.
+pub fn plan_project(mut project: Project) -> Project {
+    let world_info = project.world_registry.get(&project.target_world).cloned();
+    if let Some(world_info) = world_info {
+        project.has_http_handler_export = world_info.has_http_handler_export();
+    }
+    project.component_plan = Some(crate::wasm_plan::build_component_plan(&project));
+    project
+}
+
 /// Compile a Project to Wasm bytes using the WIR pipeline.
 ///
-/// This is the WIR pipeline entry point, parallel to `Codegen::generate_wasm`.
-/// Pipeline: Project → `tir_to_wir` (`WirModule`) → `wir_emit` (Wasm bytes)
-pub fn compile_with_wir(project: &Project) -> Vec<u8> {
+/// Pipeline: planning → `build_wir_module` (`WirModule`) → `wir_emit` (Wasm bytes)
+pub fn compile_with_wir(project: Project) -> Vec<u8> {
+    // Planning (was formerly the standalone wasm_plan phase)
+    let project = plan_project(project);
+
     // Phase 1: Build WirModule from Project
-    let wir_module = build_wir_module(project);
+    let wir_module = build_wir_module(&project);
 
     // Phase 2: Emit core module bytes from WirModule
     let core_module = emit::emit_core_module(&wir_module);
@@ -30,13 +45,8 @@ pub fn compile_with_wir(project: &Project) -> Vec<u8> {
     // Phase 2.5: Validate core module (catch errors before component wrapping)
     validate_core_module(&core_module);
 
-    // Debug: always dump latest WAT for analysis
-    if let Ok(wat) = wasmprinter::print_bytes(&core_module) {
-        let _ = std::fs::write("/tmp/wir_debug_core.wat", &wat);
-    }
-
     // Phase 3: Wrap in Component Model
-    let wasm = component::build_component(project, &core_module, &wir_module);
+    let wasm = component::build_component(&project, &core_module, &wir_module);
 
     // Phase 4: Validate
     validate_wasm(&wasm);
